+++ packages/stream_chat_flutter/lib/src/message_widget.dart
@@ -858,6 +858,7 @@ class _MessageWidgetState extends State<MessageWidget>
       hostDisplayName: hostDisplayName,
       textPadding: widget.textPadding,
       messageTheme: widget.messageTheme,
+      onLinkTap: widget.onLinkTap,
     );
   }